+++ src/Git_Remote_Updater/Settings.php
@@ -165,7 +165,7 @@ class Settings {
 	public function redirect() {
 		$redirect_url = is_multisite() ? network_admin_url( 'admin.php' ) : admin_url( 'admin.php' );
 		// phpcs:ignore WordPress.Security.NonceVerification.Missing
-		$query = isset( $_POST['_wp_http_referer'] ) ? parse_url( esc_url_raw( wp_unslash( $_POST['_wp_http_referer'] ) ), PHP_URL_QUERY ) : null;
+		$query = isset( $_POST['_wp_http_referer'] ) ? parse_url( html_entity_decode( esc_url_raw( wp_unslash( $_POST['_wp_http_referer'] ) ) ), PHP_URL_QUERY ) : null;
 		parse_str( $query, $arr );
 
 		$location = add_query_arg(
@@ -191,7 +191,7 @@ class Settings {
 				'admin_enqueue_scripts',
 				function () {
 					// phpcs:ignore WordPress.WP.EnqueuedResourceParameters.NoExplicitVersion
-					wp_register_script( 'git-remote-updater-actions', plugins_url( basename( GIT_REMOTE_UPDATER_DIR ) . '/js/git-remote-updater-switcher.js' ), [], false, true );
+					wp_register_script( 'git-remote-updater-actions', plugins_url( basename( dirname( __DIR__, 2 ) ) . '/js/git-remote-updater-switcher.js' ), [], false, true );
 					wp_enqueue_script( 'git-remote-updater-actions' );
 				}
 			);